% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{friendly_pals}
\alias{friendly_pals}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 15.
}
\usage{
friendly_pals
}
\description{
Use \code{\link[wesanderson]{wes_palette}} to construct palettes of desired length.
}
\details{
Color palettes were obtained from various websites include:
https://personal.sron.nl/~pault/#sec:qualitative
https://davidmathlogic.com/colorblind/#\%23648FFF-\%23785EF0-\%23DC267F-\%23FE6100-\%23FFB000
https://bookdown.org/hneth/ds4psy/ds4psy_files/figure-html/apx-color-define-rgb-1.png
}
\keyword{datasets}
