% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_big_grid.R
\name{theme_big_grid}
\alias{theme_big_grid}
\title{Big grid theme extension for ggplot2 plots}
\usage{
theme_big_grid()
}
\description{
Big grid theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
