% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_big_simple.R
\name{theme_big_simple}
\alias{theme_big_simple}
\title{Big simple theme extension for ggplot2 plots}
\usage{
theme_big_simple()
}
\description{
Big simple theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
