% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_black.R
\name{theme_black}
\alias{theme_black}
\title{Black theme extension for ggplot2 plots}
\usage{
theme_black()
}
\description{
Black theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
\examples{
ggplot(iris, aes(Sepal.Length, Petal.Length, color=Species)) + geom_point() + 
 labs(title="Iris") + theme_black() + facet_wrap(~Species)
}
