% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blue.R
\name{theme_blue}
\alias{theme_blue}
\title{Blue theme extension for ggplot2 plots}
\usage{
theme_blue()
}
\description{
Blue theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
