% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grey.R
\name{theme_grey}
\alias{theme_grey}
\title{Grey theme extension for ggplot2 plots}
\usage{
theme_grey()
}
\description{
Grey theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
\examples{
ggplot(iris, aes(Sepal.Length, Petal.Length, color=Species)) + geom_point() + 
 labs(title="Iris") + theme_grey() + facet_wrap(~Species)
}
