% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grid.R
\name{theme_grid}
\alias{theme_grid}
\title{Grid theme extension for ggplot2 plots}
\usage{
theme_grid()
}
\description{
Grid theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
\examples{
ggplot(iris, aes(Sepal.Length, Petal.Length, color=Species)) + geom_point() + 
 labs(title="Iris") + theme_grid() + facet_wrap(~Species)
}
