% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_red.R
\name{theme_red}
\alias{theme_red}
\title{Red theme extension for ggplot2 plots}
\usage{
theme_red()
}
\description{
Red theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
\examples{
ggplot(iris, aes(Sepal.Length, Petal.Length, color=Species)) + geom_point() + 
 labs(title="Iris") + theme_red() + facet_wrap(~Species)
}
