% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{theme_simple}
\alias{theme_simple}
\title{Simple theme extension for ggplot2 plots}
\usage{
theme_simple()
}
\description{
Simple theme for ggplot2 plots. See \href{https://ggplot2.tidyverse.org/reference/theme.html}{theme()} for complete documentation.
}
